Mobile resp (#81)
* Make communtiy-battles and ai-box mobile friendly
