+++ src/app/(non-authed)/ai-box/hooks/use-is-mobile.tsx
@@ -1,0 +1,25 @@
+import { useEffect, useState } from "react";
+
+const useIsMobile = () => {
+  const [isMobile, setIsMobile] = useState<boolean>(false);
+
+  useEffect(() => {
+    // Check if window is defined
+    if (typeof window !== "undefined") {
+      const checkMobile = () => {
+        setIsMobile(window.innerWidth <= 768);
+      };
+
+      checkMobile();
+      window.addEventListener("resize", checkMobile);
+
+      return () => {
+        window.removeEventListener("resize", checkMobile);
+      };
+    }
+  }, []);
+
+  return isMobile;
+};
+
+export default useIsMobile;
