+++ src/components/common/ora-network-modal/community-battle-modal.tsx
@@ -7,6 +7,8 @@
 /* eslint-disable tailwindcss/no-custom-classname */
 import { useEffect, useState } from "react";
 import { ArrowFatDown, Confetti, CurrencyCircleDollar } from "@phosphor-icons/react";
+import { DialogClose } from "@radix-ui/react-dialog";
+import { AiOutlineClose } from "react-icons/ai";
 import { toast } from "sonner";
 
 import { Button } from "@/components/ui/button";
@@ -129,7 +131,7 @@ const OraCommunityBattlesPromptModal = ({
 
   return (
     <Dialog open={isOpen} onOpenChange={setIsOpen}>
-      <DialogTrigger asChild className="max-lg:hidden">
+      <DialogTrigger>
         {roomState === "WIN" && !aiJudgeProcessedQuery ? (
           <button
             className={cn(
@@ -152,7 +154,10 @@ const OraCommunityBattlesPromptModal = ({
         ) : null}
       </DialogTrigger>
       {isOpen && (
-        <DialogContent className="z-[100] flex flex-col gap-12 bg-black p-4 max-lg:size-full max-lg:max-w-full max-lg:rounded-none max-lg:bg-dark-900 lg:p-8">
+        <DialogContent className="z-[100] flex flex-col gap-12 bg-black p-4 max-lg:size-full max-lg:max-h-[80vh] max-lg:max-w-full max-lg:overflow-y-auto max-lg:rounded-none max-lg:bg-dark-900 lg:p-8">
+          <DialogClose>
+            <AiOutlineClose />
+          </DialogClose>
           {transactionStatus === "loading" && (
             <div className="relative flex flex-col items-center justify-center text-xl">
               <p>Transaction in progress</p>
